Fixed issue with loading objects
Issue: Loading an object with RigidBody3D. Saving an object with RigidBody3D would store the bodyID from that session, when reading the object, it would load the bodyID from the previous session, causing Jolt to think that object is valid and not create a body on Jolt's side, preventing any physics to work on said RigidBody3D object.

Fix: Reset bodyID when saving, additionally removed the line that reads the bodyID, this way Jolt can create bodies and assign a new bodyID themselves.